fix taskcontroller

+++ app/Http/Controllers/TasksController.php
@@ -81,15 +81,18 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        
-        $task = Task::find($id);
-        if ((\Auth::user()->id) != $task->user_id){
+        if (\Auth::check()) {
+            $task = Task::find($id);
+            if ((\Auth::user()->id) != $task->user_id){
+                return redirect('/');
+            }
+            
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }else {
             return redirect('/');
         }
-        
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
     }
 
     /**
@@ -101,13 +104,18 @@ class TasksController extends Controller
     public function edit($id)
     {
         
-        $task = Task::find($id);
-         if ((\Auth::user()->id) != $task->user_id){
-            return redirect('/');
+        if (!\Auth::check()) {
+            $task = Task::find($id);
+            if ((\Auth::user()->id) != $task->user_id){
+                return redirect('/');
+            }
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }else{
+             return redirect('/');
         }
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        
     }
 
     /**
@@ -140,12 +148,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        if ((\Auth::user()->id) != $task->user_id){
+        if (\Auth::check()) {
+            $task = Task::find($id);
+            if ((\Auth::user()->id) != $task->user_id){
+                return redirect('/');
+            }
+            $task->delete();
             return redirect('/');
+        }else{
+             return redirect('/');
         }
-        $task->delete();
-
-        return redirect('/');
     }
 }
